refactor: MasterUserRecord assertions tests
move into 'pkg/test/masteruserrecord/masteruserrecord_assertion_test.go'
since the tests are related to the assertions
